+++ packages/fighting-design/image-preview/src/interface.d.ts
@@ -2,6 +2,10 @@ export interface switchImageInterface {
   (type: 'next' | 'prev'): void
 }
 
+export interface optionClickTargetKey {
+  key: string
+}
+
 export interface optionClickInterface {
   ({ key }: { key: string }): void
 }