+++ packages/fighting-design/toolbar/src/toolbar.vue
@@ -39,7 +39,11 @@
   })
 
   const handleClick: b = (evt: MouseEvent): void => {
-    const node: HTMLElement | undefined = (evt.path as HTMLElement[]).find(
+    // 针对 Safari 点击事件 event.path 为 undefined 不存在做兼容
+    // https://blog.csdn.net/weixin_43388844/article/details/120002440
+    const path: HTMLElement[] =
+      evt.path || (evt.composedPath && evt.composedPath()) || []
+    const node: HTMLElement | undefined = path.find(
       (item: HTMLElement): boolean => {
         return item.className === 'f-toolbar-item'
       }