+++ start/main.ts
@@ -1,6 +1,9 @@
 import { createApp } from 'vue'
 import App from './src/App.vue'
 
+import VConsole from 'vconsole'
+new VConsole()
+
 // 开发阶段引入
 import FightingDesign from '../packages/fighting-design/index'
 import '@fighting-design/fighting-theme'