+++ start/src/App.vue
@@ -1,32 +1,23 @@
 <script lang="ts" setup>
-  const imageList2 = [
-    'https://images.unsplash.com/photo-1558979158-65a1eaa08691?ixlib=rb-1.2.1&ixid=eyJhcHBfaWQiOjEyMDd9&auto=format&fit=crop&w=1350&q=80',
-    'https://images.unsplash.com/photo-1558979158-65a1eaa08691?ixlib=rb-1.2.1&ixid=eyJhcHBfaWQiOjEyMDd9&auto=format&fit=crop&w=1350&q=80',
-    'https://images.unsplash.com/photo-1558979158-65a1eaa08691?ixlib=rb-1.2.1&ixid=eyJhcHBfaWQiOjEyMDd9&auto=format&fit=crop&w=1350&q=80',
-    'https://images.unsplash.com/photo-1558979158-65a1eaa08691?ixlib=rb-1.2.1&ixid=eyJhcHBfaWQiOjEyMDd9&auto=format&fit=crop&w=1350&q=80',
-    'https://images.unsplash.com/photo-1558979158-65a1eaa08691?ixlib=rb-1.2.1&ixid=eyJhcHBfaWQiOjEyMDd9&auto=format&fit=crop&w=1350&q=80',
-    'https://images.unsplash.com/photo-1558979158-65a1eaa08691?ixlib=rb-1.2.1&ixid=eyJhcHBfaWQiOjEyMDd9&auto=format&fit=crop&w=1350&q=80'
-  ]
-  const imageList = [
-    {
-      url: 'https://images.unsplash.com/photo-1558979158-65a1eaa08691?ixlib=rb-1.2.1&ixid=eyJhcHBfaWQiOjEyMDd9&auto=format&fit=crop&w=1350&q=80',
-      text: '图片1'
-    },
-    {
-      url: 'https://images.unsplash.com/photo-1572276596237-5db2c3e16c5d?ixlib=rb-1.2.1&ixid=eyJhcHBfaWQiOjEyMDd9&auto=format&fit=crop&w=1350&q=80'
-    },
-    {
-      url: 'https://images.unsplash.com/photo-1551009175-8a68da93d5f9?ixlib=rb-1.2.1&ixid=eyJhcHBfaWQiOjEyMDd9&auto=format&fit=crop&w=1351&q=80'
-    },
-    {
-      url: 'https://images.unsplash.com/photo-1551009175-8a68da93d5f9?ixlib=rb-1.2.1&ixid=eyJhcHBfaWQiOjEyMDd9&auto=format&fit=crop&w=1351&q=80'
-    },
-    {
-      url: 'https://images.unsplash.com/photo-1551009175-8a68da93d5f9?ixlib=rb-1.2.1&ixid=eyJhcHBfaWQiOjEyMDd9&auto=format&fit=crop&w=1351&q=80'
-    }
-  ]
+  const change = ({ key }) => {
+    console.log(key)
+  }
 </script>
 
 <template>
-  <f-expand-card round :image-list="imageList2" />
+  <f-button @click="change">123</f-button>
+  <f-toolbar @click="change">
+    <f-toolbar-item data-key="1">
+      <f-icon size="30px" icon="f-icon-default-template" />
+    </f-toolbar-item>
+    <f-toolbar-item data-key="2">
+      <f-icon size="30px" icon="f-icon-Customermanagement" />
+    </f-toolbar-item>
+    <f-toolbar-item data-key="3">
+      <f-icon size="30px" icon="f-icon-email" />
+    </f-toolbar-item>
+    <f-toolbar-item data-key="4">
+      <f-icon size="30px" icon="f-icon-editor" />
+    </f-toolbar-item>
+  </f-toolbar>
 </template>